# BGDIINF_SB-2299: Improved performance by factor 7.5x using caching and pyproj.Transformer