Some search results might call the `transform_round_geometry` several times with
the same arguments. This function is quite slow (~50ms per call) so calling it
several times can have a significant impact.
For example the following request `/rest/services/inspire/SearchServer?type=layers&searchText=wand&geometryFormat=geojson&sr=2056`, the results in the raw sphinxsearch provides 30 matches that will call `transform_round_geometry()` 30 times with the same arguments. We could change the code to call it only once but using the caching is much simpler.

Also the `pyproj.transform()` method is deprecated and slower than the new method using `pyproj.Transformer`, therefore make use of `pyproj.Transformer`. Transforming one coordinate using `Transformer` is about ~2ms quicker than with `transform()`, for multiple coordinate the time gain is even bigger. Using the `Transformer` (without caching) improved the above request time from ~500ms ! Because creating a `Transformer` is quite time consuming (~6.5ms) we put it in a method using caching as well.

Overall with this PR the request in example above running locally using sphinxsearch on PROD (Ireland) we have the following perf

```
BASE_URL=http://localhost:5000 ITERATIONS=50 python3  -m test_search_perf
Did 50 iterations with 0 errors
avg=0.222, min=0.169, max=0.287
```
In comparison without the caching and uses of Transformer we had

```
BASE_URL=http://localhost:5000 ITERATIONS=50 python3  -m test_search_perf
Did 50 iterations with 0 errors
avg=1.635, min=1.498, max=1.753
```

So overall with this PR we have a performance gain of factor ~7.3x